[computed props] Only error on multiple string literal lower bounds
Summary:
While working on D18087304, I noticed a way to get rid of ~500 spread errors. This diff changes the way we count lower bounds on a computed property to _only_ count the string literal lower bounds. Incidentally, it gave us an opportunity for better error messages as well.

The spirit behind erroring on multiple lower bounds in computed properties is to prevent things like P119703710. Taking the enum and turning it into an object type can cause a GIANT amount of errors to be emitted, which causes flow to time out.

However, things like `StrT` (not literal), `NumT`, and `AnyT` do not take a highly optimized representation (enum) and turn it into a very not optimized representation (giant union of objects). In fact, all three of those cases are places in which Flow is currently unsound. Note, those cases are hit when we `SetPropT` with a `Computed` propref

A large portion of the error diff is from things like any propagation adding a second lower bound (see the added test case). Being more relaxed about when we complain about multiple lower bounds gets rid of those kinds of errors.

In cases where there are a ton of `any` lower bounds, we are still protected by  linearity checking.

Originally, I was ok with these errors, since it was also explicitly pointing out places where people were getting implicit anys. However, it's better to tackle that problem of computed non-literal keys separately with a more targeted error message.

Reviewed By: avikchaudhuri

Differential Revision: D18089625

fbshipit-source-id: 980e058d041a46f3bef1b772d12f57ee5411c4aa