[autocomplete] Don't duplicate this in autocomplete results
Summary:
In the previous diff, we fixed autocomplete on `this` substituted types (which probably should never reach the normalizer). Now that we aren't failing there, we end up with duplicated autocomplete results.

This diff prevents that duplication by ensuring that we only emit one `this` result.

When we start using the typed AST, we can probably revert this diff.

Reviewed By: mroch

Differential Revision: D18143882

fbshipit-source-id: b7b02b1862ceabf19b7c8ba46566c8a8c0fea4a3