Fix node module resolver bug
Summary:
https://github.com/facebook/flow/issues/8156 made me aware that some users want to be able to import their code using root-relative paths. They were misusing `module.system.node.resolve_dirname`, which worked for them due to a bug, and experiencing odd behaviors. This stack of diffs demonstrates the behavior, fixes the bug (breaking the use case), and adds first-class support for this use case.

This diff fixes the bug. We kept track of which directories contain `node_modules` (or one of its aliases), but we didn't keep track of which alias is within the container. This meant that if you configured your `.flowconfig` with

```
module.system.node.resolve_dirname=node_modules
module.system.node.resolve_dirname=.
```

then any directory `path/to/container` that contained `node_modules/` would lead to us thinking `path/to/container/./` contained node modules. This allowed people to use root-relative paths, but also treated all the code as if it were inside `node_modules/`

The fix is to keep track of which aliases we saw within each container.

Reviewed By: jbrown215

Differential Revision: D18172874

fbshipit-source-id: e6851c16b811d1160f65b647d2bb123649c7e5fd