[get-def] let definitions point at themselves
Summary:
before, if you get-def on a definition itself, like the `Foo` in `class Foo {}`, we'd realize that the request was inside the result and return no results.

both the C and TypeScript language servers still return the result, which causes the cursor to jump to the beginning of the token.

seems minor, but I think it helps to be consistent and to make it obvious that the get-def was successful.

Reviewed By: dsainati1

Differential Revision: D18234465

fbshipit-source-id: 78c92287bb54eb15f3008e9c4297078d16f72aba