Don't autocomplete binding sites
Summary:
Don't fire identifier autocompletion in binding sites since we aren't in the business of suggesting what people should name their variables.

This also subsumes D18077653 : there was a bug where AUTO332 would leak into suggestions in a certain subset of autocomplete request locations. I'm pretty convinced now that those leaks happened exactly when the requested location was in a binding site, which is why this fixes that. The tests I've added to this diff address that bug (if you look at `autocomplete.exp` in V1 of this diff you can see how AUTO332 originally leaked)

Reviewed By: nmote

Differential Revision: D18090676

fbshipit-source-id: dcc6ae59b44726c538c785c8f70927cf9e0b1794