lower bound object keys should flow as a union to upper bound indexers
Summary:
Consider the following code:

```
declare function foo<TKey, TValue, TObj: {[key: TKey]: TValue}>(obj: TObj): {[key: TValue]: TKey};

foo(someLargeObject);
```

This will trigger an `ObjT ~> ObjT` flow that will flow each of the properties in `someLargeObject` to the indexer property of `TObj` in order to produce output values for `TValue` and `TKey`. Because these are generic, they are represented as tvars, and they accumulate each of the keys and values respectively of `someLargeObject` as their lower bounds. If either of these later flow to an upper bound, this will trigger flows from each of these lower bounds to that upper bound, and this can become expensive as the size of `someLargeObject` increases.

We can take advantage, however, of the fact that the keys of an object are always an enum, to optimize here. Instead of flowing each of the keys of `someLargeObject` to `TKey` individually, we can instead flow them as a union to `TKey` since tvars and unions have the same lower bound behavior. That way, whenever `TKey` receives an upper bound, we can take advantage of our enum union optimizations to potentially avoid extra work.

Reviewed By: panagosg7

Differential Revision: D18231853

fbshipit-source-id: 4659fb0b4e13e966098dcbfabd37436d9590638a