Error if we try to update (++ or -- operators) a read-only property
Summary:
Previously, we did not emit an error when updating a read-only property. We fix this by making use of `assign_member`.

Something I noted in a comment: Updating involves both reading and writing. We need to model both of these, and ensuring an arithmetic operand should use the read type, which is affected by refinements.

Reviewed By: gabelevi

Differential Revision: D18213986

fbshipit-source-id: 56abe846ff8243595485f019e58c469a47e82c16