[get-def] log when there are too many or too few types
Summary:
when get-def jumps through a type, it can fail if the type is a tvar with 0 lower bounds (nowhere to go), or multiple bounds (can't decide where to go).

this just improves the logging so we can differentiate between the two cases.

### Background

the former case happens for empty types, like if you call it on `x` in

```
{
  let x; // never assigned, so `(x: empty)`
}
```

the latter case happens when there's no annotation and multiple writes, like:

```
{
  let x; // get-def on this x
  if (true) {
    x = 1;
  } else {
    x = 2;
  }
}
```

NOTE: we jump through types to jump across `require()`, but do so for all pattern identifiers, not just the LHS of `require` variable declarators, which I think is too broad. the examples above ideally shouldn't hit this case.

Reviewed By: jbrown215

Differential Revision: D18241222

fbshipit-source-id: 619241832d8711cb00ef0eb8d01d4642c01106e8