[spreads] Allow primitive spreads
Summary:
Spreading numbers, strings, and bools all result in `{||}`. We can model that instead of saying that those aren't objects, which helps out RN with some of the patterns they like to use.

For example:
```
const styles = {...(showBlue() && {backgroundColor: 'blue'}), color: 'red'};
```
should pass type checking and is a very interesting + reasonable pattern!

object_kit handles a number of other object operations that are not spread, but I haven't verified that we have the correct behavior when we use this new rule in those operations. Until I verify those codepaths, I'm only turning this on for spreads.

Reviewed By: gabelevi

Differential Revision: D18270818

fbshipit-source-id: a4d1f4595cc9adaf8a81393f254d46457e3df1b3