[optional-chaining] Simpler (and more amenable to method calls) approach to optional chaining
Summary:
This diff changes the implementation of optional chaining to one that is (I hope) easier to understand, doesn't use mutable references, and which helps make implementing optional chained method calls much more straightforwards (the next diff on the stack).

Before, an OptionalChainT represented *everything* in an optional chain, and contained a list of opt_use_ts (a use_t with the output tvar factored out). When something non-nullish flowed to an OptionalChainT, it would kick off an imperative loop of flowing the LHS sequentially to the opt_use_ts. Building these chains is a little hacky and relies on using references to change reasons "out from under" given points in the chain. This works, but it is a little hard to understand and fragile.

This diff changes the behavior of OptionalChainT to represent just one "link" in the chain. When a type flows into an OptionalChainT, if it's nullish, it flows to a tvar in the OptionalChainT which represents short-circuiting, and if it's not nullish, it flows directly to a use_t; in other words, an OptionalChainT just filters nullish LBs into a special accumulator tvar, and passes other types through.

We build OptionalChainTs in statement.ml by recursively descending through chains, e.g. `a?.b?.c`. At each level, we produce an output that contains "non-nullish success" and "failure/short-circuit" types. The "success" type is passed upwards through the chain and used as the receiver for further operators, while the "short-circuit" type (which includes the accumulator tvars of any OptionalChainTs in the chain as lower bounds) is, if necessary, set as a lower bound to the overall type of the expression.

So, if we have
```
declare var a: ?{b?: {c: number};
var d = a?.b?.c
```
we'll generate the flows
```
?{b?: {c: number}} ~>
  OptionalChainT (success=GetPropT("b", tout=TVAR1), short-circuit=TVAR2));
TVAR1 ~>
  OptionalChainT (success=GetPropT("c", tout=TVAR3), short-circuit=TVAR4));
TVAR2 ~> TVAR_final
TVAR4 ~> TVAR_final
TVAR3 ~> TVAR_final
```
The first flow above in turn kicks off:
```
void ~> TVAR2
{b?: {c: number}} ~> GetPropT("b", tout=TVAR1)
^ not optional any more!
?{c: number} ~> TVAR1
```
and similar for the other OptionalChainT flow.

The result is that the type of `d` is `TVAR_final`, with lower bounds `number`, `void` (from `a` being optional), and `void` (from `b` being optional).

This diff *shouldn't* cause any difference in behavior, it's just cleanup and set-up for methods. However, there is a small error diff. I can't repro why the errors weren't caught with the old approach, but they're legit errors.

Reviewed By: panagosg7

Differential Revision: D18211760

fbshipit-source-id: fc263de127f88aa917046b446057b956a7ac73f8