fix flushing of parser test output
Summary:
fixes an issue where the parser test runner didn't flush stdout after printing to it, so sometimes the output wouldn't appear until the tests finished and exited.

this was an issue for me when I accidentally introduced an infinite loop in the parser, and couldn't see which test it was coming from because its progress wasn't flushed.

Reviewed By: jbrown215

Differential Revision: D18300516

fbshipit-source-id: f24a8248dff484e63b3faaf73f7e4282e83fab96