Instead of splitting Union -> MaybeP flows, filter out the good/bad types from the union
Summary:
Previously, when encountering a flow of the form `UnionT ~> PredicateT (MaybeP)` or `NotP (MaybeP)`, we would open up the union type and flow each element of the union to the predicate, which would then be flowed to a resulting tvar which would accumulate all of the union's elements as lower bounds. Tvars like this with multiple lower bounds and no upper bounds mimic the behavior of unions, but miss out on our optimizations. Instead, we can just filter out the proper types from the union in linear time, then flow the new, intact union to the tvar, giving it a single lower bound that can benefit from our existing union optimizations.

We have to skip the optimizations when the union contains tvars however. This is because of union types that can look like `VoidT | OpenT` wherein the `OpenT` will eventually resolve to `null`, like the below.
```
declare var x: ?{f : string};

var y;

if (true) {
  y = x;
}

if (y != null && y.f != null) {

}
```

Reviewed By: jbrown215

Differential Revision: D17866756

fbshipit-source-id: bd9bfcb829e0474b8da88961e48b6d5abbd0df31