Make namespaced import store type of identifier it binds
Summary:
```
import * as foo from 'bar';
            ^^^
```
this used to be the only untyped binding site in the AST. This changes it to be typed. Autocomplete collects variables' types from their binding sites, which made it not suggest namespace imports; this fixes that. I've also deleted the now unused methods in `flow_polymorphic_ast_mapper` that were for dealing with untyped patterns.

Reviewed By: panagosg7

Differential Revision: D18239754

fbshipit-source-id: 05470bf60259f83280f0af67be6ef992ee82e5f3