Flesh out check polarity visitor
Summary:
This code was a bit of a mess. I spent some time wrapping my head around this,
and here's a bunch of cleanup.

1. We're only visiting types that arise from annotations, which means that a
bunch of types will never appear. I changed the visitor to raise an exception on
these cases, to make it clear they are unexpected, not simply unhandled. I also
added some comments to justify potentially unintuitive cases.

2. I changed the patterns to be a bit more future proof. I avoided wildcards in
records like objtype and insttype, opting to fully destructure and explicitly
ignore fields.

3. I filled out some missing cases exposed by (2).

Note that EvalT and CustomFunT types will certainly appear in annotations, but
remain TODO even after this diff.

In general, I suspect that this logic should instead be interleaved with the
code that constructs these types, instead of implementing a separate, carefully
constructed visitor. This is not because of the cost of traversing types, which
I expect is negligible. Rather, it's evidently subtle to write this visitor
correctly, whereas an interleaved logic would more likely be correct by
construction.

Reviewed By: jbrown215

Differential Revision: D15764671

fbshipit-source-id: 79d5df72bd62cba962615374de34f48a04542795