Rename dependency graph variables to be more accurate under types-first
Summary:
Previously, we had `dependency_graph` which was the sig dependency graph in types-first, and the implementation dependency graph in classic. `dependency_graph` was also used in some places where the kind of dependency graph didn't matter for the operation, such as most functions in `Pure_dep_graph_operations`. In this diff, I've renamed the uses that refer to the sig dependency graph in types-first to `sig_dependency_graph`. Under classic, the implementation dependency graph is also used as the sig dependency graph. I've left the uses that are agnostic to the kind of dependency graph named `dependency_graph`.

We also had the `all_dependency_graph`, which was the implementation dependency graph in both types-first and classic. I think that `implementation_dependency_graph` is a clearer name, so I've changed it as well.

Reviewed By: gabelevi

Differential Revision: D18330010

fbshipit-source-id: 5d44a67b642e2b2ef4822b19b47536228802dd8c