[optional-chaining] Optional chaining for methods
Summary:
This diff builds support for optional method chains.

The tricky bit for optional method chains, beyond what is needed for optional chains in general, is that method calls combine both property lookup and function calling into a single flow: `X.method(args)` will generate a flow `X ~> MethodT(method, args)`, which will look up `method` in `X` and then flow the result to `CallT(args)`. With optional chaining, *either* the lookup or the call can be optional, or both. If the call is optional (`X.method?.(args)`), we need to intercede in the logic of `X ~> MethodT` in order to filter out the nullish case for the method type before it is flowed to a `CallT`.

The approach this diff takes is to add a `ChainM` method action, which represents optional chain filtering of nullish types before they're called. When a type flows into a `MethodT` with a `ChainM` method action, the method (which may be nullish) is looked up, and then flowed into an `OptionalChainT` representing the optional chaining of the call. The non-nullish components of the method type are then flowed into a `CallT` and the process proceeds as usual.

One thing that complicates this story is that `funcalltype`s contain a "this" type, which in method calls is the method's receiver (`X` in `X.method()`). Normally, this can just be generated in statement.ml by getting the type of the target of the method lookup. However, what if this is an optional access, e.g. `X?.method()`? If we naively got the type of `X` here, the "this" type of the `funcalltype` would be `?X` (assuming that `X` is an instance of class `X`). But in the `FunT ~> CallT` flow that eventually occurs, the "this" type of the `funcalltype` will flow into the "this" type of the actual function, which in the case of this method would be `X`. This would result in a type error, and it also doesn't really make sense--in order for us to have even gotten to the function call in the first place, the receiver would have to not be nullish, becuase the optional chain would have already short-circuited if it was null.

Instead, this diff delays the binding of the receiver to the `funcalltype`'s "this" type until after the nullish types have been filtered out of it. It does so by adding a little additional bit of logic to optional chaining, by (when necessary) including a tvar in the `OptionalChainT` that takes non-nullish LBs of the optional chain as its LBs--in other words, when nullish types have been filtered away, the rest of the types flow into this tvar in addition to flowing into a use_t as described in the previous diff in this stack. This tvar is used as the "this" type for the `funcalltype`.

Most of the work of setting this up happens in statement.ml, and it is a little complicated, needing to look two levels deep into the chain and considering both whether the method lookup and the method call is optional, as well as handling refinements and test hooks etc. I've tried to document it, but I'm happy to write up as much as is needed in any part that is hard to comprehend.

Reviewed By: panagosg7

Differential Revision: D18302022

fbshipit-source-id: 5d8abe863beb98268b725d3ea924fdbf00c28c06