[hack] Replace Core_kernel with Hh_prelude to forbid polymorphic equal in decl and typing
Summary:
If we are to store decls off of the OCaml heap (via ocamlrep in cachelib), we cannot use polymorphic compare or equality functions, since they will not compare on-heap and off-heap values correctly.

After the previous diff, we use monomorphic equality functions in the decl and typing directories, but we are not protected from future regressions. This diff attempts to prevent regressions by replacing instances of `open Core_kernel` with a new module, `Hh_prelude`, which includes `Core_kernel` but shadows `=` and `<>` with a value of type `[ ˋDo_not_use_polymorphic_compare ]`.

Reviewed By: 2BitSalute

Differential Revision: D17811767

fbshipit-source-id: dda75b426a28e40f0222b63a94384fcd62c8c21b