Add ALoc.id
Summary:
Instance types, opaque types, and enums use abstract locations as identifiers.
This can lead to errors when checking a file as part of a cycle where a type
with an abstract location id can be compared to a concrete location for the same
location.

We fixed these bugs by adding a `concretize_equal` function, which tries to
convert abstract locations to concrete ones before comparison based on a reverse
lookup table.

In D16469791, dsainati1 had a clever idea to perform an optimistic concrete ->
abstract conversion at construction time instead. This diff applies the same
idea to the other uses of abstract locations as identifiers.

The implementation introduces a private type in the `ALoc` module, which can only
be constructed by the `id_of_aloc` function. Thus, `ALoc.id` inhabitants are
guaranteed to be equal if they point to the same location.

I decided to use a private type instead of an abstract type because these
identifiers are frequently used as `ALoc.t` as well, mostly for debugging. This
code can use the built-in `:>` coercion instead of (a) providing a `aloc_of_id`
function of (b) providing `ALoc.id`-specific version of various functions.

My original motivation for this diff is that I wanted to build an `ALocMap` in
the context to store some information, but I ran into the issue where two
equivalent identifiers were not considered equal, due to the representation
issue described above.

Reviewed By: nmote

Differential Revision: D18354373

fbshipit-source-id: bc45a18cac997e898a12838d5319b42ac10b41ea