Ensure abstract locations are 1-1 with concrete locations
Summary:
The signature builder will sometimes generate ASTs which re-use the same
concrete location for different nodes. When this happens, the abstract locations
pass will assign different abstract locations to each occurrence of the concrete
location.

This is fine for the abstract -> concrete conversions which we normally perform,
but we also sometimes do a concrete -> abstract conversion using a reverse
table.

Since we have a bijection, we need to be actually 1-1 and onto. We were already
onto, so this adds the additional 1-1 invariant.

Reviewed By: nmote

Differential Revision: D18354375

fbshipit-source-id: f7a04e21174280a83c35e98ad266e9a3f2c4a489