Drop unnecessary polarity field from BoundT
Summary:
In D9445076 I dropped a bunch of unnecessary stuff from BoundT types. At the
time I noted that polarity was unnecessary, but used for polarity checks. This
diff changes the polarity checking logic so the field is unused everywhere.

Before, we would walk a type for any BoundT and use the BoundT's polarity. This
change passes in a map of type parameters instead. When we find a BoundT, we
look up the corresponding type param from the map and use the polarity from
there.

I also optimized callers to avoid unnecessary traversals. There are two simple
optimizations here: (1) We don't need to walk non-polymorphic types, and (2) we
don't need to continue walking types once all type params have been shadowed.

Reviewed By: panagosg7

Differential Revision: D15764672

fbshipit-source-id: 9a32cd6ee87996eafb7380939f3b6ddaf6b8c52d