[spreads][jsx] Move JSX onto the new spread semantics
Summary:
Currently, the new spread semantics are only enabled for object literals, but not objects created via JSX. This diff moves JSX onto the new spread semantics. We should expect a sizeable error diff in React props, where we used to have all the unsound behavior from the previous spread implementation.

There was one tricky note here-- JSX does not create unsealed objects if no props are passed in, so we have to specify that we should `NOT` be unsealed if no props are provided.

Reviewed By: samwgoldman, avikchaudhuri

Differential Revision: D18235863

fbshipit-source-id: ae256935212cf5210ebc902d7b0d4900c42e97ae