resolve all EvalTs in a union when the union appears as a lower bound prior to flowing to the upper bound
Summary:
The most common reason that we miss out on our optimized union checks is because we don't realize the union is optimizable at the time that we interact with it. This can happen because the union type has unresolved tvars in it, or because it has unevaluated type destructors, just to name a couple common reasons. As an example, consider the following code snippet:

```
type A = {type : "A", field : string }
type B = {nested : {type : "B", field : number }}

type Union = A | $ElementType<B, "nested">

```

This introduces new logic to force unions in lower bound positions containing unresolved eval types to resolve all their evals before we consider their upper bounds. This way, when we see the union again, we can take advantage of our existing optimizations.

For the most part, the logic here is pretty simple. When we see a union type that contains eval types, we create a `ResolveUnionT` that forwards each of its components through the flow logic that takes care of evals and such types automatically.

The only snag is that some of these unresolved types (namely tvars or typeapps) may never be resolved at all. We don't want this to hide errors that might come from incompatibility with other types in the union, so we divide the union in two, one containing all the eval types and the other containing the remaining types. Since the behavior for lower bound unions is just a for-all quantifier, we can handle these two unions as if they don't interact at all, and gain any optimization benefits on the resolved union. (In practice, large unions contain few evals in comparison to their other types. We usually miss optimizations because of unions that contain just a single eval that ruins the party for everyone.)

Reviewed By: avikchaudhuri

Differential Revision: D18146023

fbshipit-source-id: 881459a27e896af7f964299b747db72e9d0b4cf7