modernize and rename MyMap
Summary: `MyMap.ml` was an old wrapper around ocaml's map module before it had a lot of the convenient functions it has today. In the time since it was written it rotted a lot, with many duplicate and redundant functions. This trims out the functions that are no longer necessary, and renames it to `WrappedMap` to reflect that it is a wrapper around the map api and not someone's CS101 project.

Reviewed By: samwgoldman

Differential Revision: D18361372

fbshipit-source-id: fe960bbc388db26bab0b451c0639204d2086c84e