Tweak normalization of type params
Summary:
Before this diff, the type normalizer relied on the ability to look up, for any
given type parameter, the corresponding Type.typeparam record. This imposed a
curious constraint on the Typed AST, that a class's instance type needed to be
present.

The instance type is added on the Ast.Class.body node, but it's a curious place
to smuggle a type around, since it doesn't correspond to a type, per se.

It turns out that the instance type is not actually required in the first place.
Indeed, the normalizer only needs to keep track of the name and location of type
parameters in scope.

This follows from a simple intuition: type parameters are not types! In every
case, the "type" associated with these locations was `BoundT`, which is a member
of `Type.t`, but not morally a type. The only reasonable thing to print, given a
`BoundT`, is the name of the type parameter.

This diff encodes this intuition by changing the normalizer environment to
maintain only the name and location of type parameters in scope, including the
implicit "this" type parameter for classes.

Reviewed By: panagosg7

Differential Revision: D18316518

fbshipit-source-id: 93838726c29cc0fdbc439b105c7ac23f2fec7a1c