[optional-chaining] Optional chaining deletion
Summary:
The spec for optional chaining specifies that optional chains can be targets of deletion (https://github.com/tc39/proposal-optional-chaining#optional-deletion). This diff adds support for that.

This necessitates exposing the `recur_chain` implementation function that was originally scoped inside of `subscript`, with the rationale that in an expression like `delete a?.b.c`, we need the "non-nullable" output of `a?.b` which is what `recur_chain` provides, since if `a` was nullable we wouldn't get to the deletion in the first place. Likewise, if the deletion is an expression like `delete a?.b`, we pass `a` through an `OptionalChainT` that filters out the nullable components of its type before performing the deletion.

One tricky bit arises due to refinements: normally, an expression like `delete a.a` adds to the refinement environment that `a.a` has type undefined, similar to if we wrote `a.a = undefined`. On the other hand, `delete a?.a` should not add that to the environment, so it's not just a matter of treating `delete a?.a` as identical to `delete a.a` up to optional chaining. Instead, this diff adds some additional bookkeeping to prevent this refinement.

Reviewed By: panagosg7

Differential Revision: D18337198

fbshipit-source-id: f9a6751e07b77817a0713156583ab3d68238db04