[optional-chaining] Remove OptionalChainVoid and modify reasons for better error messages
Summary:
The `OptionalChainVoid` internal type is no longer needed -- we're not passing "imaginary" voids up through optional chains any more after D18211760.

This diff removes that type, and it also adds a little bit of improved reason handling. According to the spec (https://github.com/tc39/proposal-optional-chaining#base-case), both `null` and `undefined` are converted to `undefined` in optional chains. This can lead to bad error messages like this: https://flow.org/try/#0MYewdgzgLgBAhgLnmAnjAPjMBXANrmAXhgEYBmMgdgG4AoUSWAIyWgCcBLMAcwyzwLE4AfgB0cakA

The error in the program is actually correct, because in the `null` case for `a`, `a?.a` has type `undefined`-- but it has a reason description of `null`, pointing at the original `null` annotation. This diff adds a new reason desc, `RVoidedNull`, for these `VoidT`s that originate from `null`s short-circuiting optional chains; hopefully the result will be clearer error messages. The textual description of `RVoidedNull` is "undefined (result of null short-circuiting an optional chain)"

Reviewed By: jbrown215

Differential Revision: D18340917

fbshipit-source-id: 460757017d99ede21cf6be175169a94431552b80