[parser] assert identifier is not reserved before consuming it
Summary:
I'm working on improving error recovery. Previously, `identifier` called `identifier_name`, which consumes a token, and then asserts that the *IdentifierName* is a valid *Identifier*. I want to error before consuming the token so that in followup diffs, we can recover by treating it as a missing identifier, followed by a valid keyword (for example).

Example:

```
class Foo extends implements IFoo {}
//                ^^^^^^^^^^ unexpected reserved word, but identifier_name consumed it before!
//                           the errors cascade from there: unexpected IFoo (expected implements or {), etc
```

Reviewed By: samwgoldman

Differential Revision: D18337051

fbshipit-source-id: 3bef19c98ea75dfc8aeabaa7c02c07f6616e3642