Utility to query multiple locs in file in runtests.sh
Summary:
Adds a utility to make bulk command queries in a file (e.g. `type-at-pos`). Before, we
had to write the exact location that we wanted `flow type-at-pos` to run on in test.sh.
This made it rather painful to change the source code in the test folder, as we had to
update the script to keep it in sync.

`queries_in_file 'query' 'file.js' args ...` performs 'query' on all locations
(line, col) in 'file' that have a `^` in comments at location (line+1, col).

E.g.
```
type A = number;
//   ^
```
You can also pass in flags after the ^:
```
//   ^ --expand-type-aliases
```
Invoke this passing
* the query name
* the file
* any additional flags to be applied to every query in the file

E.g.
```
query_in_file "type-at-pos" "file.js" --pretty --json
```

Reviewed By: gkz

Differential Revision: D18400044

fbshipit-source-id: f7802d08f9f294caed98e3365f0b6da2cf2ce33d