[enums] Add separate location for enum body
Summary: Previously, the enum body did not have a separate location. When using the estree_translator, this resulted in the enum body node having the same range/loc as the enum declaration itself. This is both not technically accurate, and caused some issues with Prettier when formatting enums with comments. Thus, add a separate location to the enum body. It starts after the `enum E`.

Reviewed By: mroch

Differential Revision: D18386990

fbshipit-source-id: 7bf9dc77dd512bde3db201cb0674ec1c0e9c6b97