Set 5s timeout for each check job
Summary:
This is an optimization that will make types-first faster for certain types of repositories.

If you run `flow check --profile` on a types-first repository, it prints out a table that looks like

```
==========================================Init Timings=========================================
   WALL TIME            CPU TIME         RUN/READ/SEND/IDLE/DONE/GC m/GC M      SECTION
-----------------------------------------------------------------------------------------------
111.686 (100.0%)    1992.307 (100.0%)    36%   0%   3%   0%  51%   3%   5%    <Total>
  0.000 (  0.0%)       0.000 (  0.0%)     0%   0%   0%   0%   0%   0%   0%      FetchSavedState
  8.094 (  7.2%)     114.431 (  5.7%)    23%   0%  46%   1%  20%   3%   4%      Parsing
  0.281 (  0.3%)       0.281 (  0.0%)     0%   0%   0%   0%   0%   0%   0%      PackageHeap
  1.374 (  1.2%)       1.374 (  0.1%)     0%   0%   0%   0%   0%   0%   0%      InitLibs
  2.956 (  2.6%)      12.238 (  0.6%)    46%   0%  45%   0%   6%   0%   0%      CommitModules
  2.847 (  2.5%)     107.483 (  5.4%)    72%   0%   0%   0%  26%   0%   0%      ResolveRequires
  0.667 (  0.6%)       4.940 (  0.2%)    24%   0%  55%   0%  19%   0%   0%      CalcDepsTypecheck
  0.298 (  0.3%)       0.298 (  0.0%)     0%   0%   0%   0%   0%   0%   0%      FilesToInfer
  0.376 (  0.3%)       0.376 (  0.0%)     0%   0%   0%   0%   0%   0%   0%      PruneDeps
  0.306 (  0.3%)       5.703 (  0.3%)    39%   0%   0%   0%  55%   0%   5%      EnsureParsed
  0.180 (  0.2%)       0.180 (  0.0%)     0%   0%   0%   0%   0%   0%   0%      CalcDeps
 12.574 ( 11.3%)     391.028 ( 19.6%)    88%   0%   0%   1%   4%   1%   4%      Merge
  0.064 (  0.1%)       0.064 (  0.0%)     0%   0%   0%   0%   0%   0%   0%      PrintGCStats
 73.828 ( 66.1%)    1346.066 ( 67.6%)    28%   0%   0%   0%  62%   3%   5%      Check
  7.178 (  6.4%)       7.177 (  0.4%)     0%   0%   0%   0%   0%   0%   0%      CollateErrors
  0.007 (  0.0%)       0.005 (  0.0%)     0%   0%   0%   0%   0%   0%   0%      FormatErrors
  0.655 (  0.6%)       0.662 (  0.0%)    93%   0%   0%   0%   0%   6%   0%    <Unknown total>
```

In this case, we can see the `Check` step is spending 62% of worker time in the DONE phase. That basically means that we're spending a lot of time at the end of the `Check` step with most workers finished and only a few workers still working.

This is happening because a bunch of files are slow and getting bucketed together. This leads to some buckets which are much slower than all the other buckets and lasting tens of seconds after all the other buckets have finished.

Ideally, we'd know ahead of time how long it takes to check a file and then we could bucket more cleverly. But an easier method is to add a timeout for each bucket. If a bucket times out, then we return the unchecked files to the queue.

Reviewed By: samwgoldman

Differential Revision: D18329394

fbshipit-source-id: 54425e503388d0c14635c544ce46ed21c3120889