make Parsing_service_js.parse_source_file return parse errors
Summary:
`parse_source_file` and `parse_json_file` do tolerant parses when `~fail:false`, but we didn't return the errors.

end goal: I want to log when results of autocomplete, get-def, etc are based on an invalid AST, which is not a failure (so we don't want to use `~fail:true`).

Reviewed By: nmote

Differential Revision: D18400874

fbshipit-source-id: f4df56aaa76e4c0b1ac36ba1a04415a46111d03e