give EvalTs generated from source code ALoc.id ids
Summary:
Consider the following:

`a.js`:
```
const bar = require('./b.js');
export type Bar = $PropertyType<{p:string}, 'p'> => void;
(bar: Bar);
```
and
`b.js`:
```
import type {Bar} from './a.js';
declare var bar: Bar;
module.exports = bar;
```
Naively, when checking this file under types first, we would assign Bar an `EvalT` type and generate an integer id to represent it, once during the merge phase and once again when we see the type during the check phase. Because we generate this integer twice, when we check the cast on the third line of `a.js`, we believe that the two types present there are in fact different types, and fail to recognize the fact that these `EvalT` ids are the same. By using ALoc.ids instead of ints for `EvalT`s generated from source code, we can avoid this problem, since locations do not change between merge and check. Note that this is only safe at top-level scope, since otherwise we run into problems during `generate_tests` since we are checking the same code many times.

This also adds a case to the `EvalT` checking logic to short circuit the type checker when encountering flows between `EvalT`s with the same id.

Reviewed By: samwgoldman

Differential Revision: D18401191

fbshipit-source-id: 9022c6f15cc904d49336a4a4a4b862014740b935