[autocomplete] short circuit search when finding token in bindings
Summary:
we don't return completion suggestions for binding identifiers (`const x| = ...`) because this is introducing a new name, so suggesting existing names isn't helpful.

as we walk the AST, we currently just skip these nodes, which does have the effect of returning no results, but it also continues searching the rest of the AST unnecessarily and also logs a "Autocomplete token not found in AST" response.

so, let's actually check bindings for the token and bail out if we find it, so that we can log an intentionally "empty" result.

similarly restores the `Ackey` result for object keys (`({ x| }: T)` could theoretically complete the keys in `T` but we don't support this yet), to return an intentionally empty result and log it.

Reviewed By: vrama628

Differential Revision: D18505638

fbshipit-source-id: d1b15b588c71a8229a650e96f52dd990a6cd5363