[optional-chaining] Make optional chains sensitive to refinement
Summary:
This diff makes optional chain property lookups correctly have refined heap types: it makes programs like the following well-typed, up to unnecessary optional chain lints:
```
declare var a: ?{b: number};
if (a && a.b) {
  (a?.b: number);
}
```
Before this diff, optional chains did not pick up refinements, and in the above program, we would get an error because `a?.b` could be `undefined`. This diff does *not* cause optional chains to generate refinements--as before, we're not able to use `if (a?.b) { ... }` to refine anything (that's a later diff on the stack).

One important bit added in this diff is the `~allow_optional` parameter to `Refinement.key`, which specifies whether optional members can have refinement keys (normally, they cannot). If it's set to true, which it is when we're looking up refinements in the environment, `a?.b` generates the same refinement key as `a.b`, and so we find refinements on `a.b`. It's set to false when we're *setting* refinements in the heap, in which case `a?.b` is treated as non-refinable--this way, we don't accidentally refine `a.b` to be (for example) equal to undefined just because `a?.b` is. As refinements for optional chaining continues to be built, though, we may be able to remove this flag, because we'll actually generate sensible refinements from optional chains.

Reviewed By: dsainati1

Differential Revision: D18288817

fbshipit-source-id: b48288aad110bc0964b9422a438f3a0e4b0a07e8